refactored Home pages to their own folder
